refactor: refactor api usage, move pagination to shared
