+++ src/entities/character/index.ts
@@ -1,3 +1,5 @@
 export { CharacterCard } from './ui/character-card/CharacterCard';
 export { CharacterCardList } from './ui/character-card-list/CharacterCardList';
 export { CharacterDetails } from './ui/character-details/CharacterDetails';
+export { useCharacter } from './lib/useCharacter';
+export { useCharacters } from './lib/useCharacters';
