+++ src/entities/character/ui/character-details/CharacterDetails.module.scss
@@ -2,6 +2,7 @@
   position: relative;
   display: flex;
   flex-direction: column;
+  justify-content: end;
   min-height: 700px;
   height: 100%;
   width: 100%;
