--- src/shared/api/characters/hooks/index.ts
@@ -1,2 +1,0 @@
-export { useCharacters } from './useCharacters';
-export { useCharacter } from './useCharacter';
