+++ src/shared/ui/index.ts
@@ -3,3 +3,4 @@ export { Loader } from './loader';
 export { Input } from './input';
 export { Button } from './button';
 export { ErrorMessage } from './error-message';
+export { Pagination } from './pagination';
