+++ src/shared/ui/pagination/ui/Pagination.module.scss
